feat: update to snygg v2

+++ justfile
@@ -3,15 +3,17 @@ set export
 _default:
   @just --list
 
-version := `jq '.meta.version' extension.json | tr -d '"'`
+version := `whiskers extension.tera && jq '.meta.version' extension.json | tr -d '"'`
 
-# Remove the './dist' directory
+# Remove the './dist' directory and extension.json
 clean:
-  rm -rfv ./dist
+  test -d ./dist && rm -rfv ./dist || true
+  test -f extension.json && rm -fv extension.json || true
 
 # Minify JSON files and put source files into "./dist"
 build: clean
   whiskers florisboard.tera
+  whiskers extension.tera
   cat extension.json | jq -c > dist/extension.json
   for file in `ls dist/stylesheets`; do \
       cat dist/stylesheets/$file | jq -c > dist/stylesheets/minify_$file; \
@@ -20,5 +22,5 @@ build: clean
 
 # Zip "./dist" into the "./flex" format
 zip: build
-  rm -v catppuccin-{{version}}.flex || true
+  test -f catppuccin-{{version}}.flex && rm -v catppuccin-{{version}}.flex || true
   cd dist && zip -r ../catppuccin-{{version}}.flex .
